+++ java/smile/data/DataFrameFactory.java
@@ -1,0 +1,11 @@
+package smile.data;
+
+import smile.data.vector.BaseVector;
+
+public class DataFrameFactory
+{
+  public static DataFrame construct(BaseVector... columns)
+  {
+    return new DataFrameImpl(columns);
+  }
+}