+++ src/tech/ml/dataset/impl/dataset.clj
@@ -44,7 +44,6 @@
                                 new-col-name)
                         {:existing-columns existing-names
                          :column-name new-col-name})))
-
       (new-dataset
        (ds-proto/dataset-name dataset)
        metadata
@@ -121,6 +120,16 @@
                     col)))
            (new-dataset (ds-proto/dataset-name dataset) metadata))))
 
+  (select-columns-by-index [dataset num-seq]
+    (let [col-indexes (int-array (distinct num-seq))]
+      (when-not (== (count col-indexes) (count num-seq))
+        (throw (Exception. (format "Duplicate column selection detected: %s"
+                                   num-seq))))
+      (->> col-indexes
+           (map (fn [^long idx]
+                  (.get columns idx)))
+           (new-dataset (ds-proto/dataset-name dataset) metadata))))
+
 
   (supported-column-stats [dataset]
     (ds-col/supported-stats (first columns)))
@@ -181,36 +190,49 @@
 
   The return value fulfills the dataset protocols."
   ([options ds-metadata column-seq]
-   (let [column-seq (->> (ds-col/ensure-column-seq column-seq)
-                         (map-indexed (fn [idx column]
-                                        (let [cname (ds-col/column-name
-                                                     column)]
-                                          (if (and (string? cname)
-                                                   (empty? cname))
-                                            (ds-col/set-name column idx)
-                                            column)))))
-         ;;Options was dataset-name so have to keep that pathway going.
+   (let [;;Options was dataset-name so have to keep that pathway going.
          dataset-name (or (if (map? options)
                             (:dataset-name options)
                             options)
-                          "_unnamed")
-         column-seq (if (and (map? options)
-                             (:key-fn options))
-                      (let [key-fn (:key-fn options)]
-                        (->> column-seq
-                             (map #(ds-col/set-name
-                                    %
-                                    (key-fn (ds-col/column-name %))))))
-                      column-seq)]
-     (Dataset. (vec column-seq)
-               (->> column-seq
-                    (map-indexed
-                     (fn [idx col]
-                       [(ds-col/column-name col) idx]))
-                    (into {}))
-               (assoc (col-impl/->persistent-map ds-metadata)
-                      :name
-                      dataset-name))))
+                          "_unnamed")]
+     (if-not (seq column-seq)
+       (Dataset. [] {}
+                 (assoc (col-impl/->persistent-map ds-metadata)
+                        :name
+                        dataset-name))
+       (let [column-seq (->> (ds-col/ensure-column-seq column-seq)
+                             (map-indexed (fn [idx column]
+                                            (let [cname (ds-col/column-name
+                                                         column)]
+                                              (if (and (string? cname)
+                                                       (empty? cname))
+                                                (ds-col/set-name column idx)
+                                                column)))))
+             sizes (->> (map dtype/ecount column-seq)
+                        distinct)
+             column-seq (if (== (count sizes) 1)
+                          column-seq
+                          (let [max-size (long (apply max 0 sizes))]
+                            (->> column-seq
+                                 (map #(ds-col/extend-column-with-empty
+                                        % (- max-size (count %)))))))
+             column-seq (if (and (map? options)
+                                 (:key-fn options))
+                          (let [key-fn (:key-fn options)]
+                            (->> column-seq
+                                 (map #(ds-col/set-name
+                                        %
+                                        (key-fn (ds-col/column-name %))))))
+                          column-seq)]
+         (Dataset. (vec column-seq)
+                   (->> column-seq
+                        (map-indexed
+                         (fn [idx col]
+                           [(ds-col/column-name col) idx]))
+                        (into {}))
+                   (assoc (col-impl/->persistent-map ds-metadata)
+                          :name
+                          dataset-name))))))
   ([options column-seq]
    (new-dataset options {} column-seq))
   ([column-seq]